+++ LivingLab.Web/ViewModels/LogItemViewModel.cs
@@ -11,7 +11,7 @@ public class LogItemViewModel
     public double EnergyUsage { get; set; }
 
     [Display(Name = "Interval (mins)")]
-    public double Interval { get; set; }
+    public int Interval { get; set; }
     
     [Display(Name = "Logged At")]
     public DateTime LoggedDate { get; set; }
